Убрать вложенные классы в мапперах https://github.com/AlibabaRazboinik/habit_tracker/issues/132
